% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggrSml2NumVar.R
\encoding{UTF8}
\name{AggrSml2NumVar}
\alias{AggrSml2NumVar}
\title{Aggregated comparison of two numerical variables}
\usage{
AggrSml2NumVar(data, xVar, yVar, strataVar = NULL, identiske = FALSE)
}
\arguments{
\item{data}{Input data set of class data.frame.}

\item{xVar}{Name of the x variable to be compared.}

\item{yVar}{Name of the y variable to be compared.}

\item{strataVar}{Name of stratification variable. Optional.}

\item{identiske}{When TRUE, only units with value on both x and y are used in the calculations. Default FALSE.}
}
\value{
Output of AggrSml2NumVar is a data set of class data.frame. The variables in the data frame are:
 \item{strata}{The stratum (if strataVar is given, "1" otherwise)}
 \item{Antx}{The number of units with x not missing used in the aggregation}
 \item{Anty}{The number of units with y not missing used in the aggregation}
 \item{Sumx}{The total of x in the stratum}
 \item{Sumy}{The total of y in the stratum}
 \item{SumxProsAvTotx}{The stratum total for x in percent of the population total for x: (Sumx / Totx) * 100}
 \item{SumyProsAvToty}{The stratum total for y in percent of the population total for y: (Sumy / Toty) * 100}
 \item{Diff}{The difference between the stratum totales of x and y: Sumy - Sumx}
 \item{AbsDiff}{The absolute difference: |Diff|}
 \item{DiffProsAvSumx}{The difference in percent of the stratum total for x: (Diff / Sumx) * 100}
 \item{AbsDiffProsAvSumx}{The absolute value of DiffProsAvSumx: |DiffProsAvSumx|}
 \item{DiffProsAvTotx}{The difference in percent of the population total for x: (Diff / Totx) * 100}
 \item{AbsDiffProsAvTotx}{The absolute value of DiffProsAvTotx: |DiffProsAvTotx|}
}
\description{
Calculating aggregated values for two numerical variables, useful for comparison of the variables
}
\examples{
testdata <- KostraData("testdata")

# lager en grupperingsvariabel
testdata$gr <- c(rep(3, 30), rep(5, 40), rep(1, 61), rep(2, 91), rep(3, 68), rep(4, 61),
                 rep(5, 45), rep(4, 20))

AggrSml2NumVar(data = testdata, xVar = "areal_130_eier_2014", yVar = "areal_130_eier_2015",
               strataVar = "gr", identiske = FALSE)

}
\author{
Anna Mevik
}
