% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionCalculations.R
\name{AutoFormel}
\alias{AutoFormel}
\title{AutoFormel}
\usage{
AutoFormel(s, fra2til3 = TRUE, laAvhenge = FALSE)
}
\arguments{
\item{s}{Vektor av formler}

\item{fra2til3}{Dupliserer formelbegreper uten kontoklasse ved TRUE}

\item{laAvhenge}{Erstatter IKKE avhengigheter med parenteser ved TRUE}
}
\value{
Vektor av formler
}
\description{
Formler som avhenger av hverandre omskrives slik at alle begrep på høyres side refererer til verdier fra beregning uten formler.
}
\examples{
 AutoFormel(c("D:FG2:AGD13 = D:FG2:AGD13 + D:FG2:AG12","D:FG2:AGD18 = D:FG2:AGD18 + D:FG2:AGD13"))
}
