% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanseRegnskap.R
\encoding{UTF8}
\name{BalanseRegnskapBeregning}
\alias{BalanseRegnskapBeregning}
\alias{BalanseRegnskapBeregningInput}
\alias{BalanseRegnskapBeregningHierarkiALT}
\alias{BalanseRegnskapBeregningHierarki}
\alias{BalanseRegnskapBeregningHierarkiInput}
\alias{BalanseRegnskap}
\alias{BalanseRegnskapEnPeriode}
\title{BalanseRegnskap}
\usage{
BalanseRegnskapBeregning(
  data,
  kapittelhierarki,
  kapitler,
  ...,
  perioder = NULL,
  output
)

BalanseRegnskapBeregningInput(...)

BalanseRegnskapBeregningHierarkiALT(...)

BalanseRegnskapBeregningHierarki(...)

BalanseRegnskapBeregningHierarkiInput(...)

BalanseRegnskap(data, kapittelhierarki, ..., perioder = NULL)

BalanseRegnskapEnPeriode(
  data,
  kapittelhierarki,
  kombinasjoner = NULL,
  regioner = NULL,
  storkombinasjoner = NULL,
  kapitler = NULL,
  omfang = NULL,
  output = "en",
  printData = TRUE,
  lag0300 = FALSE,
  fixRegionkode = TRUE
)
}
\arguments{
\item{data}{inputdata}

\item{kapittelhierarki}{kapittelhierarki}

\item{kapitler}{kapitler som skal med i output}

\item{...}{Parametere som sendes videre til BalanseRegnskapEnPeriode}

\item{perioder}{Perioder for beregningen. Ved NULL (default) brukes alle perioder som finnes i data.}

\item{output}{kode for type output som (brukes av BalanseRegnskapBeregningHierarki og BalanseRegnskapBeregningInput)}

\item{kombinasjoner}{kombinasjoner kapitler og regnskapsomfang som skal med i output}

\item{regioner}{regioner som skal med i output. Kan også gjøre bruk av *, ?, ! og –}

\item{storkombinasjoner}{kombinasjoner kapitler,regnskapsomfang og region som skal med i output}

\item{omfang}{regnskapsomfang som skal med i output}

\item{printData}{Ved TRUE printes to første og to siste rader av alle inputdataene}

\item{lag0300}{Ved TRUE kopieres region 0301 til 0300 i inputdata}

\item{fixRegionkode}{Ved TRUE (default): Sørger for blanke i starten/slutten fjernes og at regionkoder får 4 eller 6 plasser og ledende nuller (gir warning ved endring av input)}
}
\value{
Data frame med samme variabler som data i input og med kapitler/regnskapsomfang i henhold til annen input
}
\description{
Funksjon som ligner på KostraRegnskap, men mye enklere.
}
\note{
BalanseRegnskapBeregningInput og BalanseRegnskapBeregningHierarki går via funksjonen BalanseRegnskapBeregning.
Parametere til den funksjonen trengs bortsett fra «output».
Spesielt må altså parameterne kapittelhierarki og kapitler med.
Det skal (i utgangspunktet) være en verdi på hver av disse.
Funksjonen BalanseRegnskapBeregningHierarkiALT er alternativ til BalanseRegnskapBeregningHierarki med andre variabler i output.
}
\examples{

# Ved FALSE brukes data som ligger i pakken der bare noen regioner er med.

lesDataFraS <- FALSE
if (!lesDataFraS){
  bData <- KostraData("balanseRegnskapDataPen")   # KostraData("balanseRegnskapDataPen") er alternativ der
  inputdata <- bData$data                         #    automatisk omkoding trengs (som fixRegionkode)
  hierarki <- bData$kapittelhierarki              #    Fungerer like bra, men med flere warning
} else {

  dataPath <- "S:/Prosjekt/2228 KOSTRA publisering/Metode/Regnskap/R_balanse"
  # dataPath <- 'C:\\Users\\oyl\\Documents\\work\\Rworking\\regnskap\\R_balanse'

  hierarkiFil <- paste(dataPath, "/Kapittel_hierarki_2016_bal_KLASS", ".csv", sep = "")
  dataFil <- paste(dataPath, "/Inputdata_balanse_2016", ".csv", sep = "")
  hierarki <- read.csv2(hierarkiFil, colClasses = c("character"))
  inputdata <- read.csv2(dataFil, colClasses = c("integer", "character", "character", "character", "character", "integer"))
}

z <- BalanseRegnskap(inputdata, hierarki, regioner = c("2021", "2022", "0301", "1100"))
xA <- BalanseRegnskapBeregningInput(inputdata, hierarki, regioner = "0301", omfang = "A", kapitler = "KG3")
xB <- BalanseRegnskapBeregningInput(inputdata, hierarki, regioner = "0301", omfang = "B", kapitler = "KG3")
yA <- BalanseRegnskapBeregningHierarki(inputdata, hierarki, regioner = "0301", omfang = "A", kapitler = "KG3")
yB <- BalanseRegnskapBeregningHierarki(inputdata, hierarki, regioner = "0301", omfang = "B", kapitler = "KG3")
}
\seealso{
\code{\link{KostraRegnskap}}, \code{\link{HierarchyCompute}}, \code{\link{HierarchicalWildcardGlobbing}}
}
