% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverseFunksjoner.R
\name{CharacterReCode}
\alias{CharacterReCode}
\title{CharacterReCode}
\usage{
CharacterReCode(x, oldLevels, newLevels)
}
\arguments{
\item{x}{(character) vector}

\item{oldLevels}{oldLevels}

\item{newLevels}{newLevels}
}
\value{
character vector
}
\description{
CharacterReCode
}
\examples{
CharacterReCode(rep(1:4,2),c("2","3"),c("B","C"))

}
