% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeRegressionMulti.R
\name{CheckNumeric}
\alias{CheckNumeric}
\title{INTERNAL FUNCTION: Check values of a numeric vector or matrix according to limits.}
\usage{
CheckNumeric(
  x,
  minLimit = -Inf,
  maxLimit = Inf,
  setNA = TRUE,
  allowMissing = TRUE,
  allowInf = FALSE,
  varName = "Variable"
)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{minLimit}{lower limit}

\item{maxLimit}{upper limit}

\item{setNA}{When TRUE values outside limits are set to NA instead of error}

\item{allowMissing}{When FALSE missing values cause error}

\item{allowInf}{When FALSE -Inf and Inf always considered outside limits}

\item{varName}{String used in warning or error message}
}
\value{
x is retuned, possibly changed
}
\description{
Possibly changed data is retuned
}
\examples{
CheckNumeric(-1:10,minLimit=0,varName="Hello")
}
