% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraDiverse.R
\name{FixRegionkode}
\alias{FixRegionkode}
\title{Sørger for at regiontall blir character med 4 eller 6 plasser og ledende nuller}
\usage{
FixRegionkode(region, warningText = NULL)
}
\arguments{
\item{region}{Vektor med regionkoder. Kan være  integer, numeric, character eller factor.}

\item{warningText}{Ved ikke-NULL skrives warning med warningText dersom endring gjøres.}
}
\value{
Endrede regionkoder som en character vektor
}
\description{
Input som ikke representerer tall forblir uendret, bortsett fra eventuell omgjøring ved \code{\link{as.character}}.
}
\details{
Funksjonen kaller \code{\link{AddLeadingZeros}} med \code{places=6} for store tall (\code{>9999}) og ellers med \code{places=4}.
}
\note{
Funksjonen er ekstra grundig dokumentert siden den brukes som eksempel i workshop i bygging av  r-pakker.
}
\examples{
FixRegionkode(c(101, 301, 30107))
FixRegionkode(c("101", "301", "30107"))
FixRegionkode(c("101", "301", "30107", "Oslo"))
FixRegionkode(c("101", "301", "030107"), "Regionskoder endret")
FixRegionkode(c("0101", "0301", "030107"), "Regionskoder endret")

}
\seealso{
\code{\link{Number}}
}
\author{
Øyvind Langsrud
}
