% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverseFunksjoner.R
\name{GD}
\alias{GD}
\title{Extend an input parameter to GetData}
\usage{
GD(x, ...)
}
\arguments{
\item{x}{Original input parameter}

\item{...}{New elements to be added}
}
\value{
Extended version of x
}
\description{
Extend an input parameter to GetData
}
\examples{
GD("ola",aar=2014)
GD(list("kari",expression(aar>2014)),function(x) x^2)
}
