% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hardcode.R
\encoding{UTF8}
\name{Hardcode}
\alias{Hardcode}
\alias{TalliD}
\alias{WideAddName}
\alias{WideSep}
\alias{WideSep2}
\alias{ProtectKostraTotal}
\alias{ProtectKostraSdcStatus}
\alias{ProtectKostraSep}
\title{Small functions that return hard coded values}
\usage{
Hardcode(returnValue = "See documentation")

TalliD(returnValue = c("id", "strata", "Total"))

WideAddName(returnValue = c("", "Strata", "Total"))

WideSep(returnValue = "_")

WideSep2(returnValue = "_")

ProtectKostraTotal(returnValue = "Total")

ProtectKostraSdcStatus(returnValue = "sdcStatus")

ProtectKostraSep(returnValue = "_")
}
\value{
A vector of values
}
\description{
Small functions that return hard coded values
}
\details{
The functions are written using a parameter with a default value. This is done to make the return value visible in the documentation.
}
