% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeHistory.R
\encoding{UTF8}
\name{ImputeHistory}
\alias{ImputeHistory}
\alias{ImputeHistoryNewNames}
\alias{ImputeHistoryTall}
\alias{ImputeHistoryTallSmall}
\alias{ImputeHistoryWide}
\alias{ImputeHistoryWideSmall}
\title{Imputation of a single variable (y) by the most recent available historical value}
\usage{
ImputeHistory(
  data,
  idName = names(data)[1],
  strataName = NULL,
  xName = names(data)[3],
  yName = names(data)[4],
  weightMethod = "ordinary",
  reverse = FALSE,
  returnSameType = TRUE,
  forceIdMatching = TRUE,
  ...
)

ImputeHistoryNewNames(...)

ImputeHistoryTall(..., iD = TalliD())

ImputeHistoryTallSmall(
  ...,
  iD = TalliD(),
  keep = c("ID", "estimate", "cv", "nImputed")
)

ImputeHistoryWide(
  ...,
  addName = WideAddName(),
  sep = WideSep(),
  idNames = c("", "strata", ""),
  addLast = FALSE
)

ImputeHistoryWideSmall(
  ...,
  keep = c("id", "strata", "estimate", "cv"),
  addName = WideAddName(),
  sep = WideSep(),
  idNames = c("", "strata", ""),
  addLast = FALSE
)
}
\arguments{
\item{data}{Input data set of class data.frame}

\item{idName}{Name of id-variable(s)}

\item{strataName}{Name of starta-variable. Single strata when NULL (default)}

\item{xName}{Name of variables with historical y-value(s) (most resent first). Can be set to NULL (see yName).}

\item{yName}{Name of y-variable. When xName is NULL yName is a vector of current and historical variables (most resent first).}

\item{weightMethod}{The weight method for error calculations coded as a string: "ordinary" (default) or "ratio".}

\item{reverse}{When TRUE most resent is last instead of first (see xName and yName). Default is FALSE.}

\item{returnSameType}{When TRUE (default) and when the type of input y variable(s) is integer, the output type of
yImputed/estimate/estimateTotal is also integer. Estimates/sums are then calculated from rounded imputed values.}

\item{forceIdMatching}{When TRUE id matching in underlying GetData is forced (id as named list is not needed).}

\item{...}{Used in wrappers .... Can also be used to specify additional variable names that will be included in output (micro).}
}
\value{
Output

\strong{\code{micro}} consists of the following elements:
   \item{id}{id from input}
   \item{x}{The x variable created from input according to xName}
   \item{y}{The input y variable}
   \item{strata}{The input strata variable (can be NULL)}
   \item{category123}{Imputation groups: Not imputed (1), Imputed (3) and missing (0).
              Group 2 never happen with this function.}
   \item{yHat \emph{or estimateYHat}}{Fitted values}
   \item{yImputed \emph{or estimate}}{Imputed y-data}
   \item{rStud}{The final studentized residuals}
   \item{leaveOutResid}{The final outside-model residual}
   \item{varImputed}{Name of origin variable}

\strong{\code{aggregates}} consists of the following elements:
  \item{N}{Number of observations in each strata}
   \item{nImputed}{Number of imputed observations in each strata}
   \item{estimate}{Total estimates from imputed data}
   \item{cv}{Coefficient of variation = seEstimate/estimate}
   \item{estimateYhat}{Totale estimate based on model fits}
   \item{estimateOrig \emph{or y}}{Estimate based on original data with missing set to zero}
   \item{n}{The final number of observations in model.}
   \item{sigmaHat}{The final square root of the estimated variance parameter}
   \item{seEstimate}{The final standard error estimate of the total estimate from imputed data}

\strong{\code{total}} consists of the following elements:
   \item{Ntotal \emph{or N}}{Number of observations}
   \item{nImputedTotal \emph{or nImputed}}{Total number of imputed observations}
   \item{estimateTotal \emph{or estimate}}{Total estimate for all strata}
   \item{cvTotal or \emph{cv}}{Total cv for all strata}
}
\description{
A single x variable is created so that each element is the most resent non-missing historical value.
Missing y-values are imputed directly by this x without any model. Standard error estimates are based on
the naive model where (y-x) is assumed to be pure error.
}
\details{
This function is related to \code{\link{ImputeRegression}}
         and the structure and the names of output are similar.
         Note that missing values of x is allowed here.
         In cases were both x and y are missing a warning will occur (zero is used in total estimates).
}
\examples{

rateData <- KostraData("rateData")             # Real Kostra data set
w <- rateData$data[, c(17,19,16,5)]        # Data with id, strata, x and y

w <- w[is.finite(w[,"Ny.kostragruppe"]), ]       # Remove Longyearbyen
w[w[,"Ny.kostragruppe"]>13,"Ny.kostragruppe"]=13 # Combine small strata

# Create historical data by modifying the "original x-variable"
w2=cbind(w,x1=1.2*w[,3]*rep(c(NA,NA,1,1),107),x2=1.1*w[,3]*rep(c(NA,1),214))
ImputeHistory(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6,3)])  # Example with three historical variables - the last is complete
ImputeHistory(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6)]) # Incomplete x and a warning is produced
ImputeHistoryTall(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6,3)])
ImputeHistoryTallSmall(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6,3)])
ImputeHistoryWide(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6,3)])
ImputeHistoryWideSmall(w2, strataName = names(w2)[2], xName=names(w2)[c(5,6,3)])

# Numbers instead of names works.
# Four equivalent variants using reverse and xName=NULL
ImputeHistory(w2, strataName = 2, xName=c(5,6,3), yName=4)
ImputeHistory(w2, strataName = 2, xName=c(3,6,5), yName=4, reverse=TRUE)
ImputeHistory(w2, strataName = 2, xName=NULL, yName = c(4,5,6,3))
ImputeHistory(w2, strataName = 2, xName=NULL, yName = c(3,6,5,4), reverse=TRUE)

}
\author{
Øyvind Langsrud
}
