% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeRegressionMulti.R
\encoding{UTF8}
\name{ImputeRegressionMulti}
\alias{ImputeRegressionMulti}
\alias{ImputeRegressionMultiNewNames}
\alias{ImputeRegressionMultiTall}
\alias{ImputeRegressionMultiTallSmall}
\alias{ImputeRegressionMultiWide}
\alias{ImputeRegressionMultiWideSmall}
\title{Imputation of a several variables (y's) by a regression model using a single explanatory variable (x).}
\usage{
ImputeRegressionMulti(
  data,
  idName = names(data)[1],
  strataName = NULL,
  xName = names(data)[3],
  yNames = names(data)[4:NCOL(data)],
  ySelect = 1:length(yNames),
  methodOneComp = "mean",
  method = "ordinary",
  limitModel = 2.5,
  limitIterate = 4.5,
  limitImpute = 50,
  returnSameType = TRUE
)

ImputeRegressionMultiNewNames(...)

ImputeRegressionMultiTall(..., iD = TalliD())

ImputeRegressionMultiTallSmall(..., iD = TalliD(), keep = c("ID", "nImputed"))

ImputeRegressionMultiWide(
  ...,
  addName = WideAddName(),
  sep = WideSep(),
  idNames = c("", "strata", ""),
  addLast = FALSE
)

ImputeRegressionMultiWideSmall(
  ...,
  keep = c("id", "strata", "nImputed"),
  addName = WideAddName(),
  sep = WideSep(),
  idNames = c("", "strata", ""),
  addLast = FALSE
)
}
\arguments{
\item{data}{Input data set of class data.frame}

\item{idName}{Name of id-variable(s)}

\item{strataName}{Name of starta-variable. Single strata when NULL (default)}

\item{xName}{Name of x-variable}

\item{yNames}{Names of y-variables}

\item{ySelect}{Indices of yNames to extract a single component from}

\item{methodOneComp}{Method used to extract a single component coded as a string: "mean" (default),"pca","pcaMedian" or "pcaStd"}

\item{method}{The method (model and weight) coded as a string: "ordinary" (default), "ratio", "noconstant",
"mean" or "ratioconstant".}

\item{limitModel}{Studentized residuals limit. Above limit -> group 2.}

\item{limitIterate}{Studentized residuals limit for iterative calculation of studentized residuals.}

\item{limitImpute}{Studentized residuals limit. Above limit -> group 3.}

\item{returnSameType}{When TRUE (default) and when the type of input y variables is integer, the output type of
imputations and estimates from the final run is also integer.
Those estimates/sums are then calculated from rounded imputed values.}
}
\value{
Output of the alternative variants of the function
(Tall, Wide, Small) are constructed similar
to the variants of \code{\link{ImputeRegression}}.

Output of \strong{\code{ImputeRegressionMulti}} and \strong{\code{ImputeRegressionMultiNewNames}} (using the names after
        \code{\emph{or}}  below) is a list where the first three elements are ouput from the initial run with a single variable:
        micro has as many rows as input, aggregates has one row for each strata
        and total has a single row. The individual variables of these three elements are:

\strong{\code{micro}} consists of the following elements:
   \item{id}{id from input}
   \item{x}{The input x variable}
   \item{y}{The input y variable}
   \item{strata}{The input strata variable (can be NULL)}
   \item{category123}{The three imputation groups: representative (1), correct but not representative (2), wrong (3). }
   \item{yHat \emph{or estimateYHat}}{Fitted values}
   \item{yImputed \emph{or estimate}}{Imputed y-data}
   \item{rStud}{The final studentized residuals}
   \item{dffits}{The final DFFITS statistic}
   \item{hii}{The final leverages (diagonal elements of hat matrix)}
   \item{leaveOutResid}{The final outside-model residual}

\strong{\code{aggregates}} consists of the following elements:
   \item{N}{Number of observations in each strata}
   \item{nImputed}{Number of imputed observations in each strata}
   \item{estimate}{Total estimates from imputed data}
   \item{cv}{Coefficient of variation = seEstimate/estimate}
   \item{estimateYhat}{Totale estimate based on model fits}
   \item{estimateOrig \emph{or y}}{Estimate based on original data with missing set to zero}
   \item{coef}{The final first model coefficient}
   \item{coefB}{The final second model coefficient or zeros when only one coefficient in model.}
   \item{n}{The final number of observations in model.}
   \item{sigmaHat}{The final square root of the estimated variance parameter}
   \item{seEstimate}{The final standard error estimate of the total estimate from imputed data}
   \item{seRobust}{Robust variant of seEstimate (experimental)}

\strong{\code{total}} consists of the following elements:
   \item{Ntotal \emph{or N}}{Number of observations}
   \item{nImputedTotal \emph{or nImputed}}{Total number of imputed observations}
   \item{estimateTotal \emph{or estimate}}{Total estimate for all strata}
   \item{cvTotal or \emph{cv}}{Total cv for all strata}

   The other output elements are from the final run with all y-variables. These elements are:
   \item{\strong{MyImputed}}{Matrix of imputed y-data}
   \item{\strong{Mestimate}}{Matrix of total estimates from imputed data}
   \item{\strong{Mcv}}{Matrix of coefficient of variation = seEstimate/estimate}
   \item{\strong{MestimateTotal}}{Matrix of total estimates for all strata (a single row)}
   \item{\strong{McvTotal}}{Matrix of total cvs for all strata (a single row)}
}
\description{
Impute missing and wrong values (group 3) by the model based on representative data (group 1).
Some data are considered correct but not representative (group 2). This grouping is common for all y-variables.
}
\details{
Multivariate imputations are performed by running an imputation model within each strata.
         Thus, the three groups (category123) are the same for all y-variables.
         Division into the three groups are based on studentized residuals from a inititial run with a single variable.
         Calculations of studentized residuals
         are performed by iterativily throwing out observations from the model fitting.
         The single initial variable can be a original variable (ySelect is a single number) or a component extracted
         according to  methodOneComp.

         Calculations of studentized residuals
         are performed by iterativily throwing out observations from the model fitting.

         Missing x-values are not allowed in this version.
}
\examples{

z=KostraData("ratioTest")
z2=cbind(id=10*(1:NROW(z)),z[,c(3,1,2)],y2=z$y+z$x)
ImputeRegressionMulti(z2,strataName="k",method="ratio")
ImputeRegressionMultiNewNames(z2,strataName="k")
ImputeRegressionMultiTall(z2,strataName="k")
ImputeRegressionMultiTallSmall(z2,strataName="k")
ImputeRegressionMultiWide(z2,strataName="k")
ImputeRegressionMultiWideSmall(z2,strataName="k")

rateData <- KostraData("rateData")               # Real Kostra data set
w <- rateData$data[, c(17,19,16,5:10)]           # Data with id, strata, x and many ys
w <- w[is.finite(w[,"Ny.kostragruppe"]), ]       # Remove Longyearbyen
w[w[,"Ny.kostragruppe"]>13,"Ny.kostragruppe"]=13 # Combine small strata
ImputeRegressionMulti(w, strataName = names(w)[2])
names(w)[4:9] = paste("y",1:6,sep="")           # rename for nicer output
ImputeRegressionMulti(w, strataName = names(w)[2], method="ratio")
ImputeRegressionMulti(w, method="ratioconstant") # No strata
}
