% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySdcTableGauss.R
\name{IncDefault}
\alias{IncDefault}
\title{Default IncProgress function}
\usage{
IncDefault()
}
\description{
Default IncProgress function
}
\note{
Instead of using \code{IncProgress = IncDefault} in ProtectTable1/ProtectTable one could use \cr
\code{IncProgress = function(){cat("."); flush.console()}} \cr
but this results in wrong 
"usage line" in the documentation since ";" is not included.
}
\examples{
 IncDefault() 
}
\keyword{internal}
