% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverseFunksjoner.R
\name{IntegerDataFrame}
\alias{IntegerDataFrame}
\title{Change numeric variables in data frame to integer}
\usage{
IntegerDataFrame(x, makeWarning = FALSE, allNumeric = TRUE)
}
\arguments{
\item{x}{data frame}

\item{makeWarning}{When TRUE warning made when not all numeric variables could be changed}

\item{allNumeric}{When TRUE no change is made unless all could be changed}
}
\value{
data frame
}
\description{
Only done when the integer values are exact the same as the original values
}
