% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraData.R
\encoding{UTF8}
\name{KostraData}
\alias{KostraData}
\title{Return a data set}
\usage{
KostraData(dataset, path = NULL)
}
\arguments{
\item{dataset}{Name of data set within the Kostra package  (rateData, testdata)}

\item{path}{When non-NULL the data set is read from "path/dataset.RData"}
}
\value{
The data set
}
\description{
Return a data set
}
\examples{
 z  <- KostraData("balanseRegnskapDataPen")
 z  <- KostraData("balanseRegnskapData")
 z  <- KostraData("kostraRegnskapDataPen")
 z  <- KostraData("kostraRegnskapData")
 z  <- KostraData("rateData")
 b1 <- KostraData("testdata")
 tull <- KostraData("tulledata")

}
