% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraRegnskap.R
\encoding{UTF8}
\name{KostraRegnskap1}
\alias{KostraRegnskap1}
\title{KostraRegnskap}
\usage{
KostraRegnskap1(
  data,
  funksjonshierarki,
  artshierarki,
  kombinasjoner = NULL,
  regioner = NULL,
  inputInOutput = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{funksjonshierarki}{funksjonshierarki}

\item{artshierarki}{artshierarki}

\item{kombinasjoner}{kombinasjoner}

\item{regioner}{regioner}
}
\value{
output fra HierarchyCompute
}
\description{
Intern funksjon som kan bruke parametere i HierarchyCompute
}
\keyword{internal}
