% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraRegnskapRegionAggregering.R
\name{KostraRegnskapRegionAggregering}
\alias{KostraRegnskapRegionAggregering}
\title{Aggregering av regioner basert på output fra \code{\link{KostraRegnskap}}}
\usage{
KostraRegnskapRegionAggregering(
  data,
  hierarki,
  slettInput0 = FALSE,
  slettOutput0 = FALSE,
  fixRegionkode = TRUE,
  region = "region",
  belop = "belop",
  dimvar = c("regnskapsomfang", "kontoklasse", "art", "funksjon"),
  drop = TRUE,
  verbose = TRUE,
  highspeed = TRUE
)
}
\arguments{
\item{data}{data frame med output fra KostraRegnskap}

\item{hierarki}{aggregeringshierarki for region}

\item{slettInput0}{Ved TRUE fjernes rader med belop=0 i input}

\item{slettOutput0}{Ved TRUE fjernes rader med belop=0 i output}

\item{fixRegionkode}{Ved TRUE (default): Sørger for blanke i starten/slutten fjernes og at regionkoder (from) i hierarki får 4 eller 6 plasser og ledende nuller (gir warning ved endring av input)}

\item{region}{Variabelnavn for region}

\item{belop}{Variabelnavn for belop}

\item{dimvar}{De andre klassifiseringsvariablene ("regnskapsomfang", "kontoklasse", "art", "funksjon")}

\item{drop}{Ved TRUE begrenses output for hvert aggregat til  dimvar-kombinasjoner som har rader i inputdata som bidrar.
Ved FALSE genereres alle kombinasjoner av alle dimvar-variabler.
Ved NA vil output for alle aggregater ha alle dimvar-kombinasjoner som finnes i inputdata.}

\item{verbose}{Ved TRUE printes informasjon om fremdrift}

\item{highspeed}{Ved TRUE brukes en raskere algoritme (men ikke når drop=FALSE)}
}
\value{
En data frame
}
\description{
Funksjonen generaliserer summering ved \code{\link{aggregate}} til
 å bruke hierarki istedenfor en variabel (kostragruppe) i datasettet
}
\details{
Parameteren \code{slettOutput0} har mening i tillegg til \code{slettInput0} dersom det er negative fortegn i hierarki slik at 0-er kan oppnås ved differanser.
}
\note{
Kjernefunksjonalitet i funksjonen kan på sikt ende opp som en ny funksjon i pakka SSBtools
}
\examples{
hierarki <- Work:::Get_litehierarki()
data22 <- Work:::Get_bev_basis_R_36()[1:22, ]
KostraRegnskapRegionAggregering(data22, hierarki)
KostraRegnskapRegionAggregering(data22, hierarki, drop = NA)
KostraRegnskapRegionAggregering(data22, hierarki, drop = FALSE)
KostraRegnskapRegionAggregering(data22, hierarki, slettInput0 = TRUE, drop = NA)
}
