% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRowHierarchyCompute.R
\name{MultiRowHierarchyComputations}
\alias{MultiRowHierarchyComputations}
\title{MultiRowHierarchyComputations}
\usage{
MultiRowHierarchyComputations(
  a,
  drop0 = TRUE,
  valueName = "values",
  indName = "ind",
  ind = NULL
)
}
\arguments{
\item{a}{a}

\item{drop0}{see code}

\item{valueName}{Parameter to Stack}

\item{indName}{Parameter to Stack}

\item{ind}{see code}

\item{doStack}{Stacked output when TRUE}
}
\value{
Data frame where the "weights in weighted sum" is named as sign. Firs row (before stacking) contains the "results".
}
\description{
Showing computations behind the scenes in HierarchyCompute
}
\examples{
# Data and hierarchies used in the examples
x <- SSBtoolsData("sprt_emp")  # Employment in sport in thousand persons from Eurostat database
geoHier <- SSBtoolsData("sprt_emp_geoHier")
ageHier <- SSBtoolsData("sprt_emp_ageHier")

# Create rowSelect
rowSelect = data.frame(geo = "Europe", age = c("Y15-64", "Y15-29"), year="2015")

# Create input
a <- HierarchyCompute(x, list(age = ageHier, geo = geoHier, year = "rowFactor"), "ths_per", rowSelect = rowSelect, output = "matrixComponents")

MultiRowHierarchyComputations(a)

}
