% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LmImpute.r
\name{NaToZero}
\alias{NaToZero}
\title{INTERNAL FUNCTION: Replace missing values by zero}
\usage{
NaToZero(x, warn = NULL)
}
\arguments{
\item{x}{Input data}

\item{warn}{Warning text when non-null}
}
\value{
Input data where missing is replaced by zero
}
\description{
INTERNAL FUNCTION: Replace missing values by zero
}
\examples{
NaToZero(c(1,2,NA,4))

}
