% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeRegressionMulti.R
\name{OneComp}
\alias{OneComp}
\title{INTERNAL FUNCTION: Extract a single component from multivariate data}
\usage{
OneComp(
  data,
  variables = 1:dim(data)[2],
  method = c("mean", "pca", "pcaMedian", "pcaStd")
)
}
\arguments{
\item{data}{Data frame or matrix containing the variables}

\item{variables}{Indices of variables to extract component from}

\item{method}{Method coded as a string: "mean" (default),"pca","pcaMedian" or "pcaStd" where:
"mean" means rowMeans.
"pcaMedian" means pca/svd on data scaled by median of absolute values.
"pcaStd" means pca/svd on data centered and scaled by st.dev.}
}
\value{
A single component is retuned as a vector. This component is scaled to be a weighted
        row-sum of input so that sum(abs(weights))=1.
}
\description{
Rows with any missing elements give corresponding missing values in output.
}
\examples{
 OneComp(matrix(rnorm(15),5,3),method="pca")
}
