% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierRegression.R
\encoding{UTF8}
\name{OutlierRegression}
\alias{OutlierRegression}
\alias{OutlierRegressionMicro}
\alias{OutlierRegressionTall}
\alias{OutlierRegressionWide}
\title{Finding outliers of a sigle variable (y) by a regression model using a single explanatory variable (x).}
\usage{
OutlierRegression(
  data,
  idName = names(data)[1],
  strataName = NULL,
  xName = names(data)[3],
  yName = names(data)[4],
  method = "ordinary",
  limitModel = 2.5,
  limitIterate = 4.5
)

OutlierRegressionMicro(...)

OutlierRegressionTall(..., iD = TalliD())

OutlierRegressionWide(
  ...,
  addName = WideAddName(),
  sep = WideSep(),
  idNames = c("", "strata", ""),
  addLast = FALSE
)
}
\arguments{
\item{data}{Input data set of class data.frame}

\item{idName}{Name of id-variable(s)}

\item{strataName}{Name of starta-variable. Single strata when NULL (default)}

\item{xName}{Name of x-variable}

\item{yName}{Name of y-variable}

\item{method}{The method (model and weight) coded as a string: "ordinary" (default), "ratio", "noconstant",
"mean" or "ratioconstant".}

\item{limitModel}{Studentized residuals limit. Above limit -> outlier.}

\item{limitIterate}{Studentized residuals limit for iterative calculation of studentized residuals.}
}
\value{
Output of \strong{OutlierRegression} is a list of two data frames.  The micro data frame has as many rows as input and aggregates data frame has one row for each strata. The individual variables are:

\strong{\code{micro}} consists of the following elements:
   \item{id}{id from input}
   \item{x}{The input x variable}
   \item{y}{The input y variable}
   \item{strata}{The input strata variable (can be NULL)}
   \item{outlier}{Dummy variable: outlier (1) or not (0).}
   \item{category123}{The three imputation groups: representative (1), correct but not representative (2), wrong (3). }
   \item{yHat}{Fitted values}
   \item{rStud}{The studentized residuals from last iteration}
   \item{dffits}{The DFFITS statistic from last iteration}
   \item{hii}{The leverages (diagonal elements of hat matrix) from last iteration}
   \item{leaveOutResid}{The outside-model residual from last iteration}
   \item{limLo}{-limitModel}
   \item{limUp}{limitModel}

\strong{\code{aggregates}} consists of the following elements:
   \item{N}{Number of observations in each strata}
   \item{coef}{The final first model coefficient}
   \item{coefB}{The final second model coefficient or zeros when only one coefficient in model.}
   \item{nModel}{The final number of observations in model.}
   \item{sigmaHat}{The final square root of the estimated variance parameter}

   Output of \strong{OutlierRegressionMicro} is the single data frame micro above.

   Output of \strong{OutlierRegressionTall} and \strong{OutlierRegressionWide} are similiar to the
   functions in \code{\link{ImputeRegression}}.
}
\description{
outliers are found by using a limit for studentized residuals.
}
\details{
This function is related to \code{\link{ImputeRegression}}
         and the structure and the names of output are very similar.
         Note that missing values of x are allowed here.
}
\examples{

z = cbind(id=1:34,KostraData("ratioTest")[,c(3,1,2)])
OutlierRegression(z,strataName="k")
OutlierRegressionMicro(z,strataName="k")
OutlierRegressionTall(z,strataName="k")
OutlierRegressionWide(z,strataName="k")

rateData <- KostraData("rateData")               # Real Kostra data set
w <- rateData$data[, c(17,19,16,5)]              # Data with id, strata, x and y
w <- w[is.finite(w[,"Ny.kostragruppe"]), ]       # Remove Longyearbyen
w[w[,"Ny.kostragruppe"]>13,"Ny.kostragruppe"]=13 # Combine small strata
OutlierRegression(w, strataName = names(w)[2], method="ratio")

}
\author{
Øyvind Langsrud
}
