% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraDiverse.R
\name{PrintHeadTail}
\alias{PrintHeadTail}
\title{PrintHeadTail}
\usage{
PrintHeadTail(x, n = 2, title = NULL)
}
\arguments{
\item{x}{data.frame}

\item{n}{n}

\item{title}{title}
}
\value{
None (invisible NULL)
}
\description{
PrintHeadTail
}
