% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank2NumVar.R
\encoding{UTF8}
\name{Rank2NumVar}
\alias{Rank2NumVar}
\title{Comparing the biggest units with respect to two numerical variables}
\usage{
Rank2NumVar(
  data,
  idVar,
  xVar,
  yVar,
  strataVar = NULL,
  antall = 5,
  grense = NULL,
  identiske = FALSE
)
}
\arguments{
\item{data}{Input data set of class data.frame.}

\item{idVar}{Name of an identification variable.}

\item{xVar}{Name of the x variable to be compared.}

\item{yVar}{Name of the y variable to be compared.}

\item{strataVar}{Name of stratification variable. Optional. If strataVar is given, the calculation and listing is
performed within each stratum.}

\item{antall}{Parameter specifying how many of the biggest units to be listed. Default 5.}

\item{grense}{Parameter specifying a threshold for the units to be listed. This parameter overrules antall. Optional.}

\item{identiske}{When TRUE, only units with value on both x and y are used in the calculations. Default FALSE.}
}
\value{
Output of Rank2NumVar is a data set of class data.frame. The variables in the data frame are:
   \item{id}{The input identification variable}
   \item{x}{The input x variable}
   \item{y}{The input y variable}
   \item{strata}{The input strata variable if strataVar is given, "1" otherwise}
   \item{forh}{The ratio between x and y: y/x}
   \item{xRank}{The rank of x}
   \item{yRank}{The rank of y}
   \item{xProsAvSumx}{x in percent of the total/stratum total for x}
   \item{yProsAvSumy}{y in percent of the total/stratum total for y}
}
\description{
Calculating rank and share for two numerical variables, and the ratio between the variables
Listing big units, either the k biggest units or units with value greater than a threshold
}
\examples{
testdata <- KostraData("testdata")

# lager en grupperingsvariabel
testdata$gr <- as.character(c(rep(3, 30), rep(5, 40), rep(1, 61), rep(2, 91), rep(3, 68),
                              rep(4, 61), rep(5, 45),  rep(4, 20)))

# uten strata
Rank2NumVar(data = testdata, idVar = "Region", xVar = "areal_130_eier_2014", yVar = "areal_130_eier_2015",
            strataVar = NULL, antall = 10, grense = NULL, identiske = FALSE)

# med strata
Rank2NumVar(data = testdata, idVar = "Region", xVar = "areal_130_eier_2014", yVar = "areal_130_eier_2015",
            strataVar = "gr", antall = 10, grense = NULL, identiske = FALSE)

# med identiske = TRUE
Rank2NumVar(data = testdata, idVar = "Region", xVar = "areal_130_eier_2014", yVar = "areal_130_eier_2015",
            strataVar = "gr", antall = 10, grense = NULL, identiske = TRUE)

# med grense (overstyrer antall)
Rank2NumVar(data = testdata, idVar = "Region", xVar = "areal_130_eier_2014", yVar = "areal_130_eier_2015",
            strataVar = "gr", antall = 10, grense = 10000, identiske = FALSE)

}
\author{
Anna Mevik
}
