% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverseFunksjoner.R
\name{Separate}
\alias{Separate}
\title{Assign named elements of a list or data.frame to caller's environment}
\usage{
Separate(x)
}
\arguments{
\item{x}{A list or data.frame}
}
\value{
The function returns only NULL, but breaks the rules for ordinary functions.
        Objects are written to caller's environment.
}
\description{
Assign named elements of a list or data.frame to caller's environment
}
\examples{
Separate(data.frame(aAa=1:3, bBb=100))
aAa*bBb
}
\keyword{internal}
