% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraWrappers.R
\encoding{UTF8}
\name{SimpleHb}
\alias{SimpleHb}
\title{SimpleHb}
\usage{
SimpleHb(
  data,
  idName,
  periodName,
  measureName,
  period,
  periodCompare,
  pU,
  pA,
  pC,
  strataName
)
}
\arguments{
\item{data}{Input of Hb is a data set of class data.frame.}

\item{idName}{Name of the identification variable.}

\item{periodName}{Name of the period identification variable.}

\item{measureName}{Name of the variable to edit.}

\item{period}{value of period to edit.}

\item{periodCompare}{value of period to edit against.}

\item{pU}{Parameter that adjusts for different level of the variables.Default value 0.5.}

\item{pA}{Parameter that adjusts for small differences between the median and the 1st or 3rd quartile. Default value 0.05.}

\item{pC}{Parameter that controls the length of the confidence interval.Default value 20.}

\item{strataName}{Name of stratification variable. Optional. If strataName is given, the HB method is performed within each stratum.}
}
\value{
see Hb.r
}
\description{
SimpleHb
}
