% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraWrappers.R
\encoding{UTF8}
\name{SimpleImputeHistoryWide}
\alias{SimpleImputeHistoryWide}
\title{SimpleImputeHistoryWide}
\usage{
SimpleImputeHistoryWide(
  data,
  idName,
  periodName,
  measureName,
  strataName = NULL,
  weightMethod,
  reverse
)
}
\arguments{
\item{data}{Input data set of class data.frame}

\item{idName}{Name of id-variable}

\item{periodName}{Name of the period identification variable.}

\item{measureName}{Name of the variable to estimate.}

\item{strataName}{Name of starta-variable. Single strata when NULL (default)}

\item{weightMethod}{The weight method for error calculations coded as a string: "ordinary" (default) or "ratio".}

\item{reverse}{When TRUE most resent is last instead of first (see xName and yName). Default is FALSE.}
}
\description{
SimpleImputeHistoryWide
}
