% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraWrappers.R
\encoding{UTF8}
\name{SimpleImputeRegressionWide}
\alias{SimpleImputeRegressionWide}
\title{SimpleImputeRegressionWide}
\usage{
SimpleImputeRegressionWide(
  data,
  id,
  period,
  periodValue,
  measureName,
  compareName,
  strata,
  methodName = "ratio",
  limitModel,
  limitIterate,
  limitImpute
)
}
\arguments{
\item{data}{Input data set of class data.frame}

\item{period}{value of period to edit.}

\item{measureName}{Name of the variable to edit.}

\item{compareName}{Name of the variable to edit against.}

\item{methodName}{The method (model and weight) coded as a string: "ordinary" (default), "ratio", "noconstant",
"mean" or "ratioconstant".}

\item{limitModel}{Studentized residuals limit. Above limit -> group 2.}

\item{limitIterate}{Studentized residuals limit for iterative calculation of studentized residuals.}

\item{limitImpute}{Studentized residuals limit. Above limit -> group 3.}

\item{idName}{Name of id-variable}

\item{periodName}{Name of the period identification variable.}

\item{strataName}{Name of starta-variable. Single strata when NULL (default)}
}
\value{
see ImputeRegression.R
}
\description{
SimpleImputeRegressionWide
}
\note{
Øyvind: Denne funksjonen i Kostra-pakken opprinnelig skrevet av Håkon ser ikke ut til å være i bruk.
Det brukes kode utenfor. Har endret funksjonen til å være som siste emailutveksling med Bjørn-Andre
Har ikke oppdatert dokumentasjonen ellers.
}
