% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraWrappers.R
\encoding{UTF8}
\name{SimpleRank2NumVar}
\alias{SimpleRank2NumVar}
\title{SimpleRank2NumVar}
\usage{
SimpleRank2NumVar(
  data,
  idName,
  measureName,
  periodName,
  period,
  periodCompare,
  strataName,
  antall,
  grense,
  identiske
)
}
\arguments{
\item{data}{Input data set of class data.frame.}

\item{idName}{Name of the identification variable.}

\item{measureName}{Name of the variable to edit.}

\item{periodName}{Name of the period identification variable.}

\item{period}{value of period to edit.}

\item{antall}{Parameter specifying how many of the biggest units to be listed. Default 5.}

\item{grense}{Parameter specifying a threshold for the units to be listed. This parameter overrules antall. Optional.}

\item{identiske}{When TRUE, only units with value on both x and y are used in the calculations. Default FALSE.}

\item{xVar}{Name of the x variable to be compared.}

\item{yVar}{Name of the y variable to be compared.}

\item{strataVar}{Name of stratification variable. Optional. If strataVar is given, the calculation and listing is
performed within each stratum.}
}
\description{
SimpleRank2NumVar
}
