% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRowHierarchyCompute.R
\name{SingleRowHierarchyComputations}
\alias{SingleRowHierarchyComputations}
\title{SingleRowHierarchyComputations}
\usage{
SingleRowHierarchyComputations(
  a,
  doStack = FALSE,
  valueName = "values",
  indName = "ind"
)
}
\arguments{
\item{a}{a}

\item{valueName}{Parameter to Stack}

\item{indName}{Parameter to Stack}
}
\value{
Data frame where the "weights in weighted sum" is named as sign. Firs row (before stacking) contains the "results".
}
\description{
Showing computations behind the scenes in HierarchyCompute
}
\examples{
# Data and hierarchies used in the examples
x <- SSBtoolsData("sprt_emp")  # Employment in sport in thousand persons from Eurostat database
geoHier <- SSBtoolsData("sprt_emp_geoHier")
ageHier <- SSBtoolsData("sprt_emp_ageHier")

# Create rowSelect since a single row only
rowSelect <- data.frame(age = "Y15-64", geo = "Europe", stringsAsFactors = FALSE)

# Create input
a <- HierarchyCompute(x, list(age = ageHier, geo = geoHier, year = "colFactor"), "ths_per", rowSelect = rowSelect, output = "matrixComponents")


SingleRowHierarchyComputations(a)

SingleRowHierarchyComputations(a, TRUE)
}
