% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraDiverse.R
\name{Srn}
\alias{Srn}
\title{SortRows with rownames removed}
\usage{
Srn(x)
}
\arguments{
\item{x}{as input to SortRows}
}
\value{
sorted version of x
}
\description{
SortRows with rownames removed
}
\keyword{internal}
