% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataApply.R
\name{StrataApply}
\alias{StrataApply}
\title{INTERNAL FUNCTION: Apply a function with list output over data subsets and structure output into various matrices.}
\usage{
StrataApply(
  data,
  by,
  Fun,
  copyVar = NULL,
  byName = NULL,
  byNameDefault = "strata",
  multiUnlist = FALSE,
  FunTotal = NULL,
  FunMultiTotal = NULL,
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{by}{Separate vector or Variable in data (name or number) defining the subsets (strata)}

\item{Fun}{Function to be applied on data subsets}

\item{copyVar}{Variables from input data to be directly copied to output}

\item{byName}{Name to be used in output of variable defining subsets}

\item{byNameDefault}{Name to be used when byName=NULL and when name cannot be taken from input variable}

\item{multiUnlist}{When TRUE output elements "multi*" are split as separate elements.}

\item{FunTotal}{A function to be applied on output element "aggregates" so that  output element "total" is produced}

\item{FunMultiTotal}{A function to be applied on output element "multiAggregates" so that  output element "multiTotal" is produced}

\item{stringsAsFactors}{Option used within the function (default is FALSE).}

\item{...}{Further arguments passed to Fun}
}
\value{
Output is a list with elements:
  \item{micro}{Matrix with as many rows as data - composed of vector output.}
  \item{aggregates}{Matrix with as many rows as the number of subsets - composed of scalar output.}
  \item{total}{Result of FunTotal (typically a matrix with a single row).}
  \item{multiMicro}{List of matrices with as many rows as data - composed of matrix output.}
  \item{multiAggregates}{List of matrices with as many rows as the number of subsets - composed of single row matrix output.}
  \item{multiTotal}{Result of FunMultiTotal (typically list of matrices with a single row).}
  \item{other}{List of output that did not fit into other elements.}
  Empty elements are removed from output. Elements may also be changed due to multiUnlist.
}
\description{
INTERNAL FUNCTION: Apply a function with list output over data subsets and structure output into various matrices.
}
\examples{
 FunFun = function(data,x){
 list(max=max(data[,x]),plus100=data[,x]+100,range=t(range(data[,x])),
      xsqrtx5=cbind(data[,x],sqrt(data[,x]),5),onefive=t(1:5))}
 z = data.frame(x=(1:10)^2,k=c(1,2,2,3,3,3,4,4,4,4),six=rep(6,10))
 FunFun(z,"x")
 StrataApply(z,"k",FunFun,copyVar="six",x="x")
 StrataApply(z,"k",FunFun,copyVar="six",x="x",multiUnlist=TRUE)
}
