% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverseFunksjoner.R
\name{unAsIs}
\alias{unAsIs}
\title{Getting rid of 'AsIs' class attribute}
\usage{
unAsIs(X)
}
\arguments{
\item{X}{}
}
\value{
X without 'AsIs'
}
\description{
Getting rid of 'AsIs' class attribute
}
\details{
Notice that I take care to preserve any other class attributes that the vector might have.
}
\note{
The function is taken from
      \url{http://stackoverflow.com/questions/12865218/getting-rid-of-asis-class-attribute}
}
\examples{
class(I(TRUE))
class(unAsIs(I(TRUE)))
class(data.frame(x=I(matrix(1:12,3,4)))$x)
class(unAsIs(data.frame(x=I(matrix(1:12,3,4)))$x))
}
\author{
Verena Haunschmid
}
\keyword{internal}
